cc template working

+++ kakebo_app/forms.py
@@ -3,7 +3,7 @@ from .models import Expense
 from .constants import CURRENCY_CODES
 
 class CurrencyConversionForm(forms.Form):
-    amount = forms.DecimalField(max_digits=9, decimal_places=2)
+    amount = forms.FloatField()
     base_currency = forms.ChoiceField(choices=[(code, code) for code in CURRENCY_CODES])
     target_currency = forms.ChoiceField(choices=[(code, code) for code in CURRENCY_CODES])
